tidy up, ecju queries finished, notifications for compliance completed
