+++ templates/compliance/compliance/site_case.html
@@ -28,33 +28,33 @@
 			</a>
 			<a href="{% url 'compliance:compliance_site_details' id 'ecju-queries' %}" class="lite-tabs__tab {% if tab == 'ecju-queries' %}lite-tabs__tab--selected{% endif %}" id="link-case-notes">
 				ECJU queries
-				{% if exporter_user_notification_count.ecju_query %}
+				{% if exporter_user_notification_count.ecjuquery %}
 					<div class="lite-notification-bubble">
-						{{ exporter_user_notification_count.ecju_query }}
+						{{ exporter_user_notification_count.ecjuquery }}
 					</div>
 				{% endif %}
 			</a>
 			<a href="{% url 'compliance:compliance_site_details' id 'visit-reports' %}" class="lite-tabs__tab {% if tab == 'visit-reports' %}lite-tabs__tab--selected{% endif %}" id="link-case-notes">
 				Visits
-				{% if exporter_user_notification_count.visits %}
+				{% if exporter_user_notification_count.visitreport %}
 					<div class="lite-notification-bubble">
-						{{ exporter_user_notification_count.ecju_query }}
+						{{ exporter_user_notification_count.visitreport }}
 					</div>
 				{% endif %}
 			</a>
 			<a href="{% url 'compliance:compliance_site_details' id 'case-notes' %}" class="lite-tabs__tab {% if tab == 'case-notes' %}lite-tabs__tab--selected{% endif %}" id="link-case-notes">
 				Notes
-				{% if exporter_user_notification_count.case_notes %}
+				{% if exporter_user_notification_count.casenote %}
 					<div class="lite-notification-bubble">
-						{{ exporter_user_notification_count.ecju_query }}
+						{{ exporter_user_notification_count.casenote }}
 					</div>
 				{% endif %}
 			</a>
 			<a href="{% url 'compliance:compliance_site_details' id 'generated-documents' %}" class="lite-tabs__tab {% if tab == 'generated-documents' %}lite-tabs__tab--selected{% endif %}" id="link-case-notes">
 				Generated Documents
-				{% if exporter_user_notification_count.generated_documents %}
+				{% if exporter_user_notification_count.generatedcasedocument %}
 					<div class="lite-notification-bubble">
-						{{ exporter_user_notification_count.ecju_query }}
+						{{ exporter_user_notification_count.generatedcasedocument }}
 					</div>
 				{% endif %}
 			</a>
