+++ templates/compliance/compliance/tabs/visits.html
@@ -14,6 +14,7 @@
 						<a class="govuk-link govuk-link--no-visited-state app-icon-label" href="{% url 'compliance:compliance_visit_details' id visit_report.id "ecju-queries" %}">
 							{{ visit_report.reference_code }}
 						</a>
+						{% if visit_report.exporter_user_notification_count %}<span class="lite-notification-bubble" id="eua-notifications"><span class="govuk-visually-hidden"> (</span>{{ visit_report.exporter_user_notification_count }}<span class="govuk-visually-hidden"> notifications)</span></span>{% endif %}
 					</td>
 					<td class="govuk-table__cell">
 						{{ visit_report.visit_date|default:"Not set" }}
