+++ templates/includes/ecju-queries.html
@@ -44,7 +44,7 @@
 			<div class="app-ecju-query__item app-ecju-query__item--right">
 				<p class="app-ecju-query__heading">
 					<a href="{% url "organisation:members:user" ecju_query.responded_by_user.id %}" class="govuk-link govuk-link--no-visited-state">
-						{{ ecju_query.responded_by_user.first_name }} {{ ecju_query.responded_by_user.last_name }}
+						{{ ecju_query.responded_by_user.name }}
 					</a>
 				</p>
 				<div class="app-ecju-query__text">
